Simplified how the t variable works.
Instead of the user being able to control the rate at which t is incremented, t is just a float representing the number of seconds since drawing started. More specifically:

t = (currentTimeInMilliseconds - drawStartTimeInMilliseconds) / 1000.0

This makes the interface a little bit nicer/cleaner, because chances are the user always wants the drawing to run as fast as possible on their system, but this way they can still change the rate at which t affects their drawing by multiplying t by a factor, and how this factor affects the drawing will be consistent on different computers, since it's just based on the number of seconds elapsed.

However, implementing t this way kind of changes some of the really cool drawings that are based on bitwise operators, like (x^x + y^y)^t. You no longer get to see what the drawing is for every integer value of t, because t is no longer an integer, so it's not as good for drawings that work best with integers. That's why I put this in a separate branch instead of master, I'm not sure if this is the best way of doing it.